[FIX] Fixing id changing in persistence mock

+++ src/app/features/clients/utils/persistence.mock.ts
@@ -44,7 +44,7 @@ export class PersistenceMock {
     ]);
 
     localStorage.setItem('clientsStored', updatedClientsInStorage);
-    localStorage.setItem('currentId', JSON.stringify(newId));
+    if (replaceId) localStorage.setItem('currentId', JSON.stringify(newId));
   }
 
   public static deleteClient(clientId: number) {
